# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library Book Tracker System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  student: "student",
  librarian: "librarian",
  admin: "admin",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role?: RegisterBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  student: "student",
  librarian: "librarian",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type BookBranch = (typeof BookBranch)[keyof typeof BookBranch];

export const BookBranch = {
  CSE: "CSE",
  IT: "IT",
  Civil: "Civil",
  Mechanical: "Mechanical",
  Electrical: "Electrical",
} as const;

export interface Book {
  id: number;
  title: string;
  author: string;
  subject: string;
  branch: BookBranch;
  availability: boolean;
  section: string;
  rackNumber: string;
  rowNumber: string;
  shelfNumber: string;
  isbn?: string | null;
  createdAt?: string;
}

export type BookWithBorrowCount = Book & {
  borrowCount: number;
};

export type CreateBookBodyBranch =
  (typeof CreateBookBodyBranch)[keyof typeof CreateBookBodyBranch];

export const CreateBookBodyBranch = {
  CSE: "CSE",
  IT: "IT",
  Civil: "Civil",
  Mechanical: "Mechanical",
  Electrical: "Electrical",
} as const;

export interface CreateBookBody {
  title: string;
  author: string;
  subject: string;
  branch: CreateBookBodyBranch;
  availability?: boolean;
  section: string;
  rackNumber: string;
  rowNumber: string;
  shelfNumber: string;
  isbn?: string | null;
}

export type UpdateBookBodyBranch =
  (typeof UpdateBookBodyBranch)[keyof typeof UpdateBookBodyBranch];

export const UpdateBookBodyBranch = {
  CSE: "CSE",
  IT: "IT",
  Civil: "Civil",
  Mechanical: "Mechanical",
  Electrical: "Electrical",
} as const;

export interface UpdateBookBody {
  title?: string;
  author?: string;
  subject?: string;
  branch?: UpdateBookBodyBranch;
  availability?: boolean;
  section?: string;
  rackNumber?: string;
  rowNumber?: string;
  shelfNumber?: string;
  isbn?: string | null;
}

export interface BulkUploadBody {
  /** Raw CSV content */
  csvData: string;
}

export interface BulkUploadResponse {
  inserted: number;
  failed: number;
  errors: string[];
  total: number;
}

export interface BooksListResponse {
  books: Book[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface BranchStat {
  branch: string;
  total: number;
  available: number;
  borrowed: number;
}

export interface BooksStatsResponse {
  branches: BranchStat[];
  totalBooks: number;
  availableBooks: number;
  borrowedBooks: number;
}

export interface BorrowBody {
  bookId: number;
  dueDays?: number;
}

export interface ReturnBody {
  borrowId: number;
}

export interface BorrowRecord {
  id: number;
  userId: number;
  bookId: number;
  book?: Book;
  user?: User;
  borrowDate: string;
  dueDate: string;
  returnDate?: string | null;
  isOverdue: boolean;
  fineDays?: number | null;
}

export interface BorrowHistoryResponse {
  records: BorrowRecord[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface WishlistBody {
  bookId: number;
}

export interface WishlistItem {
  id: number;
  userId: number;
  bookId: number;
  book?: Book;
  createdAt?: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface ChatBody {
  message: string;
  conversationHistory?: ChatMessage[];
}

export interface ChatResponse {
  message: string;
  role: string;
}

export interface AdminDashboard {
  totalBooks: number;
  totalUsers: number;
  activeBorrows: number;
  overdueBorrows: number;
  totalBorrowsAllTime: number;
  branchStats: BranchStat[];
  mostBorrowedBooks: BookWithBorrowCount[];
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  borrow: "borrow",
  return: "return",
  register: "register",
  book_added: "book_added",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  userId?: number;
  userName?: string;
  bookTitle?: string | null;
  createdAt: string;
}

export type GetBooksParams = {
  page?: number;
  limit?: number;
  search?: string;
  branch?: string;
  subject?: string;
  availability?: boolean;
  author?: string;
};

export type GetPopularBooksParams = {
  branch?: string;
  limit?: number;
};

export type GetBorrowHistoryParams = {
  page?: number;
  limit?: number;
};

export type GetRecentActivityParams = {
  limit?: number;
};