# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Library Book Tracker System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /books:
    get:
      operationId: getBooks
      tags: [books]
      summary: Get books with pagination and filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: branch
          in: query
          schema:
            type: string
        - name: subject
          in: query
          schema:
            type: string
        - name: availability
          in: query
          schema:
            type: boolean
        - name: author
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated list of books
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BooksListResponse"
    post:
      operationId: createBook
      tags: [books]
      summary: Create a new book
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookBody"
      responses:
        "201":
          description: Book created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
  /books/bulk-upload:
    post:
      operationId: bulkUploadBooks
      tags: [books]
      summary: Bulk upload books via CSV
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkUploadBody"
      responses:
        "200":
          description: Bulk upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkUploadResponse"
...
  /books/popular:
    get:
      operationId: getPopularBooks
      tags: [books]
      summary: Get popular books (most borrowed)
      parameters:
        - name: branch
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Popular books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookWithBorrowCount"
  /books/{id}:
    get:
      operationId: getBook
      tags: [books]
      summary: Get a book by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateBook
      tags: [books]
      summary: Update a book
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookBody"
      responses:
        "200":
          description: Updated book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
    delete:
      operationId: deleteBook
      tags: [books]
      summary: Delete a book
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [student, librarian, admin]
          default: student
      required:
        - name
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, librarian, admin]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        subject:
          type: string
        branch:
          type: string
          enum: [CSE, IT, Civil, Mechanical, Electrical]
        availability:
          type: boolean
        section:
          type: string
        rackNumber:
          type: string
        rowNumber:
          type: string
        shelfNumber:
          type: string
        isbn:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - author
        - subject
        - branch
        - availability
        - section
        - rackNumber
        - rowNumber
        - shelfNumber
    BookWithBorrowCount:
      allOf:
        - $ref: "#/components/schemas/Book"
        - type: object
          properties:
            borrowCount:
              type: integer
          required:
            - borrowCount
    CreateBookBody:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        subject:
          type: string
        branch:
          type: string
          enum: [CSE, IT, Civil, Mechanical, Electrical]
        availability:
          type: boolean
          default: true
        section:
          type: string
        rackNumber:
          type: string
        rowNumber:
          type: string
        shelfNumber:
          type: string
        isbn:
          type: string
          nullable: true
      required:
        - title
        - author
        - subject
        - branch
        - section
        - rackNumber
        - rowNumber
        - shelfNumber
    UpdateBookBody:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        subject:
          type: string
        branch:
          type: string
          enum: [CSE, IT, Civil, Mechanical, Electrical]
        availability:
          type: boolean
        section:
          type: string
        rackNumber:
          type: string
        rowNumber:
          type: string
        shelfNumber:
          type: string
        isbn:
          type: string
          nullable: true
    BulkUploadBody:
      type: object
      properties:
        csvData:
          type: string
          description: Raw CSV content
      required:
        - csvData
    BulkUploadResponse:
      type: object
      properties:
        inserted:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            type: string
        total:
          type: integer
      required:
        - inserted
        - failed
        - errors
        - total
    BooksListResponse:
      type: object
      properties:
        books:
          type: array
          items:
            $ref: "#/components/schemas/Book"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - books
        - total
        - page
        - limit
        - totalPages
...