# lib/api-zod/src/generated/types/booksListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library Book Tracker System API
 * OpenAPI spec version: 0.1.0
 */
import type { Book } from "./book";

export interface BooksListResponse {
  books: Book[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}