# lib/api-zod/src/generated/types/bulkUploadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library Book Tracker System API
 * OpenAPI spec version: 0.1.0
 */

export interface BulkUploadBody {
  /** Raw CSV content */
  csvData: string;
}