# lib/api-zod/src/generated/types/updateBookBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library Book Tracker System API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateBookBodyBranch } from "./updateBookBodyBranch";

export interface UpdateBookBody {
  title?: string;
  author?: string;
  subject?: string;
  branch?: UpdateBookBodyBranch;
  availability?: boolean;
  section?: string;
  rackNumber?: string;
  rowNumber?: string;
  shelfNumber?: string;
  isbn?: string | null;
}